M05.3b: load base SDXL with target dtype to avoid mixed-precision inference error

+++ src/rd_lora/eval/generate.py
@@ -90,7 +90,7 @@ def load_eval_pipeline(
         raise ImportError("diffusers.StableDiffusionXLPipeline is required for eval image generation")
 
     dtype = _resolve_torch_dtype(torch_dtype)
-    pipe = StableDiffusionXLPipeline.from_pretrained(base_model_id)
+    pipe = StableDiffusionXLPipeline.from_pretrained(base_model_id, torch_dtype=dtype)
     try:
         pipe.load_lora_weights(str(checkpoint_path), weight_name="pytorch_lora_weights.safetensors")
     except TypeError:
